backend/requirements.txt
Django==5.2.12
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.2
django-import-export==4.4.0
drf-spectacular==0.29.0
python-decouple==3.8
gunicorn==25.3.0
whitenoise==6.12.0
pillow==11.3.0
mysqlclient==2.2.7
